gio: Fix various implicit conversions from size_t to smaller types
Basically various trivial instances of the following MSVC compiler
warning:
```
../gio/gio-tool-set.c(50): warning C4267: '=': conversion from 'size_t' to 'int', possible loss of data
```

+++ gio/gapplicationcommandline.c
@@ -666,8 +666,8 @@ const gchar *
 g_application_command_line_getenv (GApplicationCommandLine *cmdline,
                                    const gchar             *name)
 {
-  gint length = strlen (name);
-  gint i;
+  size_t length = strlen (name);
+  size_t i;
 
   /* TODO: expand on windows */
   if (cmdline->priv->environ)
